requirements.txt
alabaster==0.7.12
asn1crypto==0.24.0
astroid==2.2.5
autopep8==1.4.4
Babel==2.6.0
boto3==1.9.134
botocore==1.12.134
certifi==2019.3.9
cffi==1.12.3
chardet==3.0.4
Click==7.0
coverage==4.5.3
cryptography==2.7
DBUtils==1.3
docutils==0.14
Flask==1.0.2
Flask-Script==2.0.6
gunicorn==19.9.0
idna==2.8
imagesize==1.1.0
isort==4.3.17
itsdangerous==1.1.0
Jinja2==2.10.1
jmespath==0.9.4
jwt==0.6.1
lazy-object-proxy==1.3.1
lxml==4.4.1
Mako==1.0.9
MarkupSafe==1.1.1
mccabe==0.6.1
Naked==0.1.31
packaging==19.0
pycodestyle==2.5.0
pycparser==2.19
pycrypto==2.6.1
Pygments==2.3.1
PyJWT==1.7.1
pylint==2.3.1
PyMySQL==0.9.3
pyparsing==2.4.0
python-dateutil==2.8.0
pytz==2019.1
PyYAML==5.1.2
requests==2.21.0
s3transfer==0.2.0
selenium==3.141.0
shellescape==3.4.1
six==1.12.0
snowballstemmer==1.2.1
Sphinx==2.0.1
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp==1.0.2
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.3
typed-ast==1.3.4
urllib3==1.24.2
uuid==1.30
weixin-python==0.5.7
Werkzeug==0.15.2
wrapt==1.11.1
